fixed errors, added participant data to match.py
